+++ src/config.js
@@ -4,4 +4,3 @@
  * @private
  */
 exports.DEBUG_ENABLED = false;
-