+++ src/packets/GetDbParameterPacket.js
@@ -27,11 +27,7 @@ function GetDbParameterPacket(options) {
  * @param writer
  */
 GetDbParameterPacket.prototype.write = function (writer) {
-  var bufferLength = DATA_TYPES.DATA_LENGTH_SIZEOF + DATA_TYPES.CAS_INFO_SIZE +
-    DATA_TYPES.BYTE_SIZEOF +
-    DATA_TYPES.INT_SIZEOF + DATA_TYPES.INT_SIZEOF;
-
-  writer._writeInt(bufferLength - DATA_TYPES.DATA_LENGTH_SIZEOF - DATA_TYPES.CAS_INFO_SIZE);
+  writer._writeInt(this.getBufferLength() - DATA_TYPES.DATA_LENGTH_SIZEOF - DATA_TYPES.CAS_INFO_SIZE);
   writer._writeBytes(DATA_TYPES.CAS_INFO_SIZE, this.casInfo);
   writer._writeByte(CAS.CASFunctionCode.CAS_FC_GET_DB_PARAMETER);
   writer._writeInt(DATA_TYPES.INT_SIZEOF);
@@ -61,3 +57,10 @@ GetDbParameterPacket.prototype.parse = function (parser) {
   return this;
 };
 
+GetDbParameterPacket.prototype.getBufferLength = function () {
+	var bufferLength = DATA_TYPES.DATA_LENGTH_SIZEOF + DATA_TYPES.CAS_INFO_SIZE +
+			DATA_TYPES.BYTE_SIZEOF +
+			DATA_TYPES.INT_SIZEOF + DATA_TYPES.INT_SIZEOF;
+
+	return bufferLength;
+};