+++ test/test_BasicMultiFetch_Sequence.js
@@ -73,11 +73,11 @@ exports['test_BasicMultiFetch_Sequence'] = function (test) {
       client.close(cb);
     }
   ], function (err) {
-    if (err === null) {
+    if (err) {
+      throw err;
+    } else {
       Helpers.logInfo('Test passed.');
       test.done();
-    } else {
-      throw 'Error executing test!';
     }
   });
 };