+++ test/test_Blob_Insert.js
@@ -58,11 +58,11 @@ exports['test_Blob_Insert'] = function (test) {
     }
   ],
   function (err) {
-    if (err === null) {
+    if (err) {
+      throw err;
+    } else {
       Helpers.logInfo('Test passed.');
       test.done();
-    } else {
-      throw err.message;
     }
   });
 };