+++ test/test_Encoding_Russian.js
@@ -35,12 +35,12 @@ exports['test_Encoding_Russian'] = function (test) {
       client.close(cb);
     }
   ], function (err) {
-    if (err === null) {
+    if (err) {
+      throw err;
+    } else {
       Helpers.logInfo('Connection closed.');
       Helpers.logInfo('Test passed.');
       test.done();
-    } else {
-      throw err.message;
     }
   });
 };